update: refactoring
