+++ src/components/Common/BaseTextareaCounter/BaseTextareaCounter.vue
