+++ src/components/TheHeader/TheDropdownSettings/TheDropdownSettings.vue
@@ -40,7 +40,7 @@
 
 <script>
 import BaseIcon from "../../Common/BaseIcon/BaseIcon.vue";
-import BaseTooltip from "../../Common/BaseTooltip.vue";
+import BaseTooltip from "../../Common/BaseTooltip/BaseTooltip.vue";
 import TheDropdownSettingsMain from "./TheDropdownSettingsMain/TheDropdownSettingsMain.vue";
 import TheDropdownSettingsAppearance from "./TheDropdownSettingsAppearance/TheDropdownSettingsAppearance.vue";
 import TheDropdownSettingsLanguages from "./TheDropdownSettingsLanguages/TheDropdownSettingsLanguages.vue";
