+++ src/components/TheVideos/VideoItem/VideoItemInfo/VideoItemInfo.vue
@@ -26,7 +26,7 @@
 
 import BaseIcon from "../../../Common/BaseIcon/BaseIcon.vue";
 import VideoItemDropdown from "./VideoItemDropdown/VideoItemDropdown.vue";
-import BaseTooltip from "../../../Common/BaseTooltip.vue";
+import BaseTooltip from "../../../Common/BaseTooltip/BaseTooltip.vue";
 
 export default {
   components: {
